refactor: create file for veryfile email and add utils function for user auth
