+++ src/components/Signup/resend-code.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const ResendCode = () => {
+  return <div>ResendCode</div>;
+};
+
+export default ResendCode;
