+++ src/utils/auth.js
@@ -1,3 +1,15 @@
+export function saveToken(token) {
+  localStorage.setItem("token", JSON.stringify(token));
+}
+
+export function saveUser(userObj) {
+  localStorage.setItem("user", JSON.stringify(userObj));
+}
+
+export function getToken() {
+  return JSON.parse(localStorage.getItem("token"));
+}
+
 export function isLoggedIn() {
   const token = JSON.parse(localStorage.getItem("token"));
   const user = JSON.parse(localStorage.getItem("user"));
